phase 4 m4.10: rename packages/executor → packages/scheduler
Mechanical rename. The package's role evolved from "polls tasks and
spawns CLIs inline" to "claims null-runtime pending sessions as the
in-process fallback claimant when no daemon is bound to the agent."
The new name reflects the scheduling/coordination character of what
remains; the spawn responsibility itself fades as agents bind to
daemons.

Changes:
- Directory: packages/executor → packages/scheduler
- Package: @beevibe/executor → @beevibe/scheduler
- Bin: beevibe-executor → beevibe-scheduler
- Env: BEEVIBE_EXECUTOR_HEALTH_PORT → BEEVIBE_SCHEDULER_HEALTH_PORT
- All imports + scripts updated

The deeper Phase 6 cleanup (gut spawn, server-fallback-mesh worker,
orphan reaper as a separate module) stays as follow-up — the rename
alone aligns naming without changing behavior.
