+++ scripts/m7-e2e.ts
@@ -322,7 +322,7 @@ async function main(): Promise<void> {
     await new Promise<void>((resolveP, reject) => {
       const child = spawn(
         "pnpm",
-        ["-r", "--filter", "@beevibe/api", "--filter", "@beevibe/executor", "build"],
+        ["-r", "--filter", "@beevibe/api", "--filter", "@beevibe/scheduler", "build"],
         { cwd: repoRoot, stdio: "inherit" },
       );
       child.on("exit", (code) =>
@@ -361,7 +361,7 @@ async function main(): Promise<void> {
     });
     exec = spawnBinary({
       tag: "exec",
-      cwd: resolve(repoRoot, "packages/executor"),
+      cwd: resolve(repoRoot, "packages/scheduler"),
       env: childEnv,
     });
     log(`  api pid=${api.pid} exec pid=${exec.pid}`);
